# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Customer 360 Intelligence Platform",
    "description": "\nCustomer 360 Intelligence Platform API.\n\nDeterministic customer data, agent-generated insights and grounded natural-language Q&A over a\nread-only banking dataset. Monetary values are integer cents internally and decimal strings at this\nboundary; every response carries a correlation ID and trace ID in `meta`.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/customers/{customer_id}/insights": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "Stream the AI insight cards as server-sent events",
        "description": "Stream every insight card (all agents except the offer recommendation) over SSE.",
        "operationId": "stream_insights_customers__customer_id__insights_get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/customers/{customer_id}/recommendations": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "Stream the AI recommendation card as server-sent events",
        "description": "Stream the next-best-offer recommendation card over SSE.",
        "operationId": "stream_recommendations_customers__customer_id__recommendations_get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/customers/{customer_id}/ask": {
      "post": {
        "tags": [
          "ask"
        ],
        "summary": "Ask a grounded natural-language question about a customer (SSE)",
        "description": "Authorize, audit, run the Q&A graph, and stream the grounded answer (task 9.5).",
        "operationId": "ask_customers__customer_id__ask_post",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ask": {
      "post": {
        "tags": [
          "ask"
        ],
        "summary": "Ask a grounded natural-language question across all entitled customers (SSE)",
        "description": "Answer a cross-customer question and stream the grounded answer (task 9.6).\n\nThe search-landing \"Ask anything\" entry point. There is no customer to authorize up front, so\nthe agent resolves one with the entitlement-scoped ``customer_search`` tool and every reading\ntool re-authorizes the id it resolves to \u2014 entitlement is enforced per result, not skipped. The\naccess is audited fail-closed before answering, and a refusal (a tool that reached a\nnon-entitled customer) is audited as a denied access, disclosing nothing.",
        "operationId": "ask_all_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AskRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "description": "The natural-language question."
          },
          "session_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Session Id",
            "description": "Conversation session; memory is keyed by (session_id, customer_id)."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question",
          "session_id"
        ],
        "title": "AskRequest",
        "description": "The Q&A request body: a question and the session it belongs to (design \u00a710.2)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}